Fix integer truncation and missing NULL checks in upb
Four hardening fixes for upb internals:

1. compare_unknown.c: Change oldsize/newsize from int to size_t to
   prevent truncation overflow when computing allocation sizes, and add
   missing NULL check after upb_grealloc.

2. map_sorter.h/map_sorter.c: Change _upb_mapsorter and _upb_sortedmap
   struct fields and _upb_mapsorter_resize parameter from int to size_t
   to match _upb_Map_Size return type and prevent same truncation
   overflow pattern as compare_unknown.c.

3. copy.c: Add missing NULL check after upb_Arena_Malloc in
   upb_Clone_StringView before memcpy. The parallel code path in
   upb_Clone_MessageValue already checks correctly.

4. array.h: Add overflow check in _upb_Array_NewMaybeAllowSlow for
   init_capacity << elem_size_lg2 to match the safe pattern used in
   _upb_Array_Realloc (which uses upb_ShlOverflow). The combined check
   guards against both shift overflow and subsequent addition overflow.

+++ upb/message/map_sorter.c
@@ -105,15 +105,15 @@ static int (*const compar[kUpb_FieldType_SizeOf])(const void*, const void*) = {
 };
 
 static bool _upb_mapsorter_resize(_upb_mapsorter* s, _upb_sortedmap* sorted,
-                                  int size) {
+                                  size_t size) {
   sorted->start = s->size;
   sorted->pos = sorted->start;
   sorted->end = sorted->start + size;
 
   if (sorted->end > s->cap) {
-    const int oldsize = s->cap * sizeof(*s->entries);
+    const size_t oldsize = s->cap * sizeof(*s->entries);
     s->cap = upb_RoundUpToPowerOfTwo(sorted->end);
-    const int newsize = s->cap * sizeof(*s->entries);
+    const size_t newsize = s->cap * sizeof(*s->entries);
     s->entries = upb_grealloc(s->entries, oldsize, newsize);
     if (!s->entries) return false;
   }
@@ -124,7 +124,7 @@ static bool _upb_mapsorter_resize(_upb_mapsorter* s, _upb_sortedmap* sorted,
 
 bool _upb_mapsorter_pushmap(_upb_mapsorter* s, upb_FieldType key_type,
                             const upb_Map* map, _upb_sortedmap* sorted) {
-  int map_size = _upb_Map_Size(map);
+  size_t map_size = _upb_Map_Size(map);
 
   if (!_upb_mapsorter_resize(s, sorted, map_size)) return false;
